+++ .gitignore
@@ -7,3 +7,5 @@ Thumbs.db
 
 /public/css
 /public/js
+
+/.idea
