+++ gulpfile.js
@@ -1,4 +1,7 @@
-var elixir = require('laravel-elixir');
+var gulp = require('gulp');
+var browserify = require('browserify');
+var babelify = require('babelify');
+var source = require('vinyl-source-stream');
 
 var scripts = [
 	'Notify.js',
@@ -12,8 +15,21 @@ var scripts = [
 	'Router.js',
 ]
 
-elixir(function(mix) {
-	mix.less('app.less')
-		.babel(scripts)
-		.browserify('all.js', null, 'public/js');
+gulp.task('scripts', function() {
+	browserify('resources/assets/js/main.js')
+		.transform(babelify.configure({ optional: ['es7.decorators', 'es7.classProperties'] }))
+		.bundle()
+		.pipe(source('all.js'))
+		.pipe(gulp.dest('public/js'))
 });
+
+gulp.task('watch', function() {
+	gulp.watch('resources/assets/js/**/*.js', ['scripts']);
+});
+
+gulp.task('default', ['watch', 'scripts']);
+
+// elixir(function(mix) {
+// 	mix.less('app.less');
+// 	mix.babel(scripts).browserify('all.js', null, 'public/js');
+// });
