+++ package.json
@@ -1,11 +1,14 @@
 {
   "private": true,
   "dependencies": {
-    "bootstrap-sass": "^3.0.0",
+    "autobind-decorator": "^1.3.2",
+    "babelify": "^6.3.0",
+    "browserify": "^11.2.0",
+    "es6-shim": "^0.33.6",
     "gulp": "^3.8.8",
     "history": "^1.9.1",
-    "laravel-elixir": "^3.0.0",
     "react": "^0.13.3",
-    "react-router": "1.0.0-rc1"
+    "react-router": "1.0.0-rc1",
+    "vinyl-source-stream": "^1.1.0"
   }
 }
