+++ resources/assets/js/ListHandler/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+import Handler from './Handler';
+
+export default Handler;
