+++ resources/assets/js/Router.js
@@ -1,20 +1,25 @@
 'use strict';
 
-let routes = (
+import React from 'react';
+import { Router, Route, IndexRoute } from 'react-router';
+import createBrowserHistory from 'history/lib/createBrowserHistory';
+
+import App from './App';
+import Pages from './Pages';
+
+export default (
 	<Router history={ createBrowserHistory() }>
 		<Route path="/" component={ App }>
-			<IndexRoute component={ Home } />
-			<Route path="news" component={ News }>
-				<Route path=":page" component={ News }/>
+			<IndexRoute component={ Pages.Home } />
+			<Route path="news" component={ Pages.News }>
+				<Route path=":page" component={ Pages.News }/>
 			</Route>
-			<Route path="list" component={ ListPortal }/>
-			<Route path="list/*" component={ List }/>
-			<Route path="u/:steamId" component={ Profile }/>
-			<Route path="search/:searchId" component={ Search }/>
-			<Route path="donate" component={ Home }/>
-			<Route path="privacy" component={ Privacy }/>
+			<Route path="list" component={ Pages.ListPortal }/>
+			<Route path="list/*" component={ Pages.List }/>
+			<Route path="u/:steamId" component={ Pages.Profile }/>
+			<Route path="search/:searchId" component={ Pages.Search }/>
+			<Route path="donate" component={ Pages.Home }/>
+			<Route path="privacy" component={ Pages.Privacy }/>
 		</Route>
 	</Router>
 );
-
-React.render(routes, document.getElementById('app'));
