+++ resources/assets/js/main.js
@@ -1,0 +1,7 @@
+'use strict';
+
+import 'es6-shim';
+import React from 'react';
+import routes from './Router';
+
+React.render(routes, document.getElementById('app'));
