+++ resources/assets/js/pages/Search.js
@@ -1,6 +1,10 @@
 'use strict';
 
-class Search extends BasicComp {
+import React from 'react';
+import BasicComp from '../BasicComp';
+import List from './List';
+
+export default class Search extends BasicComp {
 	render() {
 		return <List search={ this.props.params.searchId } />;
 	}
