+++ resources/assets/js/pages/index.js
@@ -1,0 +1,11 @@
+'use strick';
+
+import Home from './Home';
+import News from './News';
+import ListPortal from './ListPortal';
+import List from './List';
+import Profile from './Profile';
+import Search from './Search';
+import Privacy from './Privacy';
+
+export default { Home, News, ListPortal, List, Profile, Search, Privacy }
