+++ resources/assets/js/pages/listPortal.js
@@ -1,20 +1,20 @@
 'use strict';
 
-class ListPortal extends BasicComp {
-	constructor(props)
-	{
-		super(props);
+import React from 'react';
+import { Link } from 'react-router';
+import BasicComp from '../BasicComp';
 
+export default class ListPortal extends BasicComp {
+	constructor(props) {
+		super(props);
 		this.state = { my_list: [], friends_list: [] };
 	}
 
-	componentDidMount()
-	{
+	componentDidMount() {
 		this.fetchLists();
 	}
 
-	fetchLists()
-	{
+	fetchLists() {
 		this.request.fetchLists = $.ajax({
 			url: '/api/v1/list',
 			dataType: 'json',
@@ -27,15 +27,13 @@ class ListPortal extends BasicComp {
 		});
 	}
 
-	componentWillReceiveProps(props)
-	{
+	componentWillReceiveProps(props) {
 		let updatedState = $.extend({}, this.state, {my_list: props.parentState.my_list});
 
 		this.setState(updatedState);
 	}
 
-	renderMyList(data)
-	{
+	renderMyList(data) {
 		if(data.length < 1) return <div className="custom-list"></div>;
 
 		let myList = data.map((list, index) =>
@@ -78,8 +76,7 @@ class ListPortal extends BasicComp {
 		);
 	}
 
-	renderFriendsList(data)
-	{
+	renderFriendsList(data) {
 		if(data.length < 1) return <div className="custom-list"></div>;
 
 		let friendsList = data.map((list, index) =>
@@ -127,8 +124,7 @@ class ListPortal extends BasicComp {
 		);
 	}
 	
-	render()
-	{
+	render() {
 		let myList, friendsList;
 
 		myList = this.renderMyList(this.state.my_list);
