+++ resources/assets/js/pages/news.js
@@ -1,16 +1,17 @@
 'use strict';
 
-class News extends BasicComp {
-	componentDidMount()
-	{
+import React from 'react';
+import BasicComp from '../BasicComp';
+
+export default class News extends BasicComp {
+	componentDidMount() {
 		let page = this.props.params.page;
 		if(page == undefined || page < 1) page = 1;
 
 		this.fetchNews(page);
 	}
 
-	fetchNews(page)
-	{
+	fetchNews(page) {
 		this.request.fetchNews = $.ajax({
 			url: '/api/v1/news/?page='+page,
 			dataType: 'json',
@@ -29,8 +30,7 @@ class News extends BasicComp {
 		});
 	}
 
-	render()
-	{
+	render() {
 		var state, news, pagination;
 
 		state = this.state;
