+++ resources/assets/js/partials/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+import Header from './Header';
+import Footer from './Footer';
+
+export default { Header, Footer }
