+++ resources/assets/less/pages/profile.less
@@ -105,6 +105,8 @@
 	}
 
 	.profile-vacstatus {
+		padding: 0 0 15px; 
+
 		.title {
 			font-weight: 600;
 			font-size: 1.5em;
@@ -123,9 +125,6 @@
 				margin: auto;
 			}
 		}
-
-		padding: 0 0 15px;
-		font-size: 1.1em;
 	}
 
 	.table {
